+++ aldrin/src/lib.rs
@@ -69,8 +69,8 @@ mod function_call_map;
 mod handle;
 mod lifetime;
 mod object;
+mod pending_reply;
 mod promise;
-mod reply;
 mod serial_map;
 #[cfg(test)]
 mod test;
@@ -112,7 +112,7 @@ pub use event::Event;
 pub use handle::Handle;
 pub use lifetime::{Lifetime, LifetimeId, LifetimeScope};
 pub use object::Object;
+pub use pending_reply::PendingReply;
 pub use promise::Promise;
-pub use reply::Reply;
 pub use unknown_call::UnknownCall;
 pub use unknown_event::UnknownEvent;